Consolidated Gyro Files
No more _p files and existing files re-named to just gyro
